feat: added banner carousel as component with supporting res files
