+++ app/src/main/java/com/example/android/architecture/blueprints/todoapp/util/Util.kt
@@ -1,5 +1,7 @@
 package com.example.android.architecture.blueprints.todoapp.util
 
+import com.example.android.architecture.blueprints.todoapp.R
+
 object Util {
 
     fun getBannerUrlList() = listOf(
@@ -8,5 +10,13 @@ object Util {
         "https://evaboot.com/wp-content/uploads/2023/07/verify-email-linkedin-sales-navigator-1-1024x256.png",
         "https://scontent.fblr1-6.fna.fbcdn.net/v/t39.30808-6/464037491_8351590514936571_1860232605708386253_n.jpg?_nc_cat=100&ccb=1-7&_nc_sid=cc71e4&_nc_ohc=FG9hg57BqbYQ7kNvgFalVL6&_nc_zt=23&_nc_ht=scontent.fblr1-6.fna&_nc_gid=AmQnnVnlcPQzB5ERACi5WJk&oh=00_AYC0lemKkEkhIr_8Hx6F2kNxzFcZdu3-0iStv__MNF_Agg&oe=67559E61",
         "https://images.ctfassets.net/wowgx05xsdrr/1pMa0mC60V172wQ7LAQFO3/ee19e2db47431ec1a7af65cb213c0a26/article-header-social-media-marketing-bigcommerce.jpg?fm=webp&w=3840&q=75"
+    )
+
+    fun getBannerResFiles() = listOf(
+        R.drawable.banner_image_01,
+        R.drawable.banner_image_02,
+        R.drawable.banner_image_03,
+        R.drawable.banner_image_04,
+        R.drawable.banner_image_05
     )
 }
